Change region/domain to vagrant-local from global
The hashicorp nomad examples and the command-line default use the
'global' region. To match, I've been using '.global', but 'global' is a
top-level domain and should be avoided for local development. Instead,
make it clear this is a local vagrant env, and the domain should match.
The scripts have been updated to use *.vagrant.local for certificates
and /etc/hosts entries and configs.